Update HW_Main.cpp
The program gets the file and reads a line then sends it to the function RCCount which counts the characters by using the index. Once the index hits 80 it moves to the next line and after it read all the content from the file line it exits the function and repeats.